# Optimize English to words converter