--- a/src/Benchmarks/EnglishToWordsBenchmark.cs
+++ b/src/Benchmarks/EnglishToWordsBenchmark.cs
@@ -1,0 +1,13 @@
+[MemoryDiagnoser]
+public class EnglishToWordsBenchmark
+{
+    EnglishNumberToWordsConverter converter = new();
+
+    [Benchmark]
+    public string ToWords() =>
+        converter.Convert(long.MaxValue);
+
+    [Benchmark]
+    public string ToWordsOrdinal() =>
+        converter.ConvertToOrdinal(int.MaxValue);
+}