Remove recursion & reduce allocations.

Before:
```

BenchmarkDotNet v0.13.12, Windows 10 (10.0.19045.4046/22H2/2022Update)
12th Gen Intel Core i7-12800H, 1 CPU, 20 logical and 14 physical cores
.NET SDK 8.0.201
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2


```
| Method         | Mean     | Error   | StdDev   | Gen0   | Allocated |
|--------------- |---------:|--------:|---------:|-------:|----------:|
| ToWords        | 333.2 ns | 6.70 ns | 19.22 ns | 0.1316 |   1.62 KB |
| ToWordsOrdinal | 366.2 ns | 8.01 ns | 23.37 ns | 0.1349 |   1.66 KB |

After:

```

BenchmarkDotNet v0.13.12, Windows 10 (10.0.19045.4046/22H2/2022Update)
12th Gen Intel Core i7-12800H, 1 CPU, 20 logical and 14 physical cores
.NET SDK 8.0.201
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2


```
| Method         | Mean     | Error   | StdDev  | Gen0   | Allocated |
|--------------- |---------:|--------:|--------:|-------:|----------:|
| ToWords        | 136.9 ns | 2.79 ns | 8.13 ns | 0.0489 |     616 B |
| ToWordsOrdinal | 150.0 ns | 3.18 ns | 9.37 ns | 0.0522 |     656 B |